Add interactive rebase and merge steps to PR guide
Relatively often, people will submit many commits, merges, and fixes of
their previous commits in a pull request. It would almost always be
better for them to squash all these commits down into 1 or 2 commits.

To help reduce these occurances I've added a sentence encouraging the
use of interactive rebasing and linking to GitHub's overview of the
subject.

A new step in the PR guide is to remind people to merge their topic
branch into the upstream branch. This makes sure that they resolve any
merge conflicts and submit code that is against the latest version of
the remote branch.

Close #4