Move the guidelines into CONTRIBUTING.md
Based on a great suggestion by @sopel.

Moving the guidelines themselves into a CONTRIBUTING.md makes it easier
to apply them to a repository than having them directly in the README.

This change includes some minor copy changes and relaxes the language,
since the main problem is not the strength of the language but the fact
that so few people ever read the CONTRIBUTING.md file.