Adding Flowchart to help contributors.
While defining a 1-size-fit-all "flow" can be
close to impossible, I think a _generic enough_
procedure can be outlined.

This is an attempt.

The image is rendered using PlantUML (http://plantuml.sourceforge.net/),
a self-executing Java JAR/library that can generate UML
out of a defined language.
Source file for the flowchart is provided.